add gun choose, add furniture

+++ Assets/Scripts/DamageListeners.cs
@@ -9,7 +9,7 @@ public abstract class DamageListeners : MonoBehaviour
         Enemy.EnemyKilled += OnEnemyKill;
     }
 
-    protected abstract void OnDamageTaken(float damage);
+    protected abstract void OnDamageTaken(int damage);
 
     protected abstract void OnEnemyKill();
 }
