+++ Assets/Scripts/ScoresCounter.cs
@@ -5,14 +5,14 @@ using TMPro;
 public class ScoresCounter : DamageListeners
 {
     private TextMeshProUGUI scoresUI;
-    private float scores;
+    private int scores;
 
     private void Start()
     {
         scoresUI = GetComponentInChildren<TextMeshProUGUI>();
     }
 
-    protected override void OnDamageTaken(float damage)
+    protected override void OnDamageTaken(int damage)
     {
         scores += damage;
         scoresUI.text = scores.ToString();
